refactor: split appmap.unittest
Move the implementation details of unittest integration into
_appmap.unittest.
